requirements.txt
yfinance
requests
python-telegram-bot==13.7
pytz